requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-datetime-widget==0.9.3
easy-thumbnails==2.7.1
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.3.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
